abstract data generation
